refactor/use jsx for component files
