+++ client/pages/blog.jsx
