+++ client/pages/blog/[slug].jsx
